requirements.txt
asgiref==3.3.4
Django==3.2
pytz==2021.1
sqlparse==0.4.1

gunicorn==20.1.0
psycopg2-binary==2.8.5
google-cloud-secret-manager==2.4.0
google-auth==1.30.0
django-storages[google]==1.11.1
django-environ==0.4.5
pytest==6.2.5
pytest-django==4.5.2

# django-debug-toolbar